style : 액티비티 각 디렉터리로 이동
