+++ app/src/main/AndroidManifest.xml
@@ -47,8 +47,8 @@
 
         <activity android:name=".view.LoginActivity" android:theme="@style/Theme.Dangder.NoActionBar"/>
         <activity android:name=".view.MainActivity" />
-        <activity android:name=".view.SignUpActivity"/>
-        <activity android:name=".view.InitDogActivity" android:windowSoftInputMode="adjustPan"/>
+        <activity android:name=".view.signup.SignUpActivity"/>
+        <activity android:name=".view.initdog.InitDogActivity" android:windowSoftInputMode="adjustPan"/>
 
     </application>
