+++ app/src/main/java/com/viewpoint/dangder/view/IntroActivity.kt
@@ -8,6 +8,7 @@ import com.viewpoint.dangder.R
 import com.viewpoint.dangder.action.Actions
 import com.viewpoint.dangder.base.BaseActivity
 import com.viewpoint.dangder.databinding.ActivityIntroBinding
+import com.viewpoint.dangder.view.initdog.InitDogActivity
 import com.viewpoint.dangder.viewmodel.LoginViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import io.reactivex.rxjava3.kotlin.addTo
